Work on #24 select Nodes labels
Promoted tvNodes to NodeTableView in network_view.py
Added: contextMenu to the Columns Headers
Added: select_tv_nodes_label() to gui.py that handles the contextMenu
Added: display_tv_nodes_label() to gui.py that display the corresponding
data in both gvNetwork and gvTsne

Started: on_tv_selection_changed() in gui.py that selects nodes in the
graphViews if they are selected in the tv. Not working yet.